pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vehicle-transport-automation"
version = "0.2.0"
description = "Email to ClickUp/Central Dispatch pipeline for vehicle transport"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Y7 Agency"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pdfplumber>=0.10.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "openpyxl>=3.1.0",
    # API dependencies
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "httpx>=0.24.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
]

[project.scripts]
dispatch = "main:main"

[project.urls]
Homepage = "https://github.com/sergiivorotyntsev/CENTRALDISPATCH"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "extractors*", "ingest*", "models*", "services*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "E402",   # Module level import not at top of file (needed for sys.path)
    "B008",   # Function call in default argument (FastAPI/Pydantic pattern)
    "B023",   # Function definition doesn't bind loop variable (closures)
    "N806",   # Variable in function should be lowercase (constants in functions)
    "N815",   # Mixed case variable in class scope (match external APIs)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "UP035"]  # Unused imports + legacy typing
"tests/*" = ["B904"]                # Re-raise patterns less critical in tests
"scripts/*" = ["B904", "B007"]      # Re-raise/unused loop vars in scripts
"api/routes/*" = ["B904"]           # HTTPException re-raises don't need chain
"services/*" = ["B904", "B007"]     # ImportError re-raises, loop vars
"extractors/*" = ["F821"]           # Forward references to DocumentStructure
